поправки в unit-тестах
